refactor: merge only app module providers in example apps infra
